src/app/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
click==8.1.7
dnspython==2.6.1
fastapi==0.114.2
h11==0.14.0
httptools==0.6.1
idna==3.10
mongomock==4.2.0.post1
motor==3.5.1
packaging==24.1
pillow==10.4.0
pydantic==1.10.13
pydantic_core==2.23.3
pymongo==4.8.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
sentinels==1.0.0
sniffio==1.3.1
starlette==0.38.5
typing_extensions==4.12.2
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.0.1
